refactor: restructure MCP Gateway by moving start-mcp.ts logic to mcp-index.ts and organizing related files into a dedicated mcp directory
