+++ src/mcp/middleware/cacheBackend/index.ts
@@ -1,7 +1,7 @@
 import { Context } from 'hono';
 import { env } from 'hono/adapter';
 import { createMiddleware } from 'hono/factory';
-import { createCacheBackendsCF } from '../../services/cache';
+import { createCacheBackendsCF } from '../../../shared/services/cache';
 
 export const cacheBackendMiddleware = createMiddleware(
   async (c: Context, next) => {
