+++ src/shared/services/cache/types.ts
