--- src/start-mcp.ts
@@ -1,51 +1,0 @@
-#!/usr/bin/env node
-
-import { serve } from '@hono/node-server';
-
-import app from './mcp-index';
-
-// Extract the port number from the command line arguments
-const defaultPort = process.env.PORT ? parseInt(process.env.PORT) : 8788;
-const args = process.argv.slice(2);
-const portArg = args.find((arg) => arg.startsWith('--port='));
-const port = portArg ? parseInt(portArg.split('=')[1]) : defaultPort;
-
-const isHeadless = args.includes('--headless');
-
-const server = serve({
-  fetch: app.fetch,
-  port: port,
-});
-
-const url = `http://localhost:${port}`;
-
-// Loading animation function
-async function showLoadingAnimation() {
-  const frames = ['⠋', '⠙', '⠹', '⠸', '⠼', '⠴', '⠦', '⠧', '⠇', '⠏'];
-  let i = 0;
-
-  return new Promise((resolve) => {
-    const interval = setInterval(() => {
-      process.stdout.write(`\r${frames[i]} Starting MCP Gateway...`);
-      i = (i + 1) % frames.length;
-    }, 80);
-
-    // Stop after 1 second
-    setTimeout(() => {
-      clearInterval(interval);
-      process.stdout.write('\r');
-      resolve(undefined);
-    }, 500);
-  });
-}
-
-// Clear the console and show animation before main output
-console.clear();
-await showLoadingAnimation();
-
-// Main server information with minimal spacing
-console.log('\x1b[1m%s\x1b[0m', '🚀 Your MCP Gateway is running at:');
-console.log('   ' + '\x1b[1;4;32m%s\x1b[0m', `${url}`);
-
-// Single-line ready message
-console.log('\n\x1b[32m✨ Ready for connections!\x1b[0m');
